backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autogluon-tabular>=1.4.0",
    "flask==2.2.5",
    "flask-cors==3.0.10",
    "flask-opentracing==1.1.0",
    "flask-restful==0.3.10",
    "flask-testing==0.8.1",
    "grpcio==1.64.1",
    "grpcio-opentracing==1.1.4",
    "grpcio-reflection==1.48.2",
    "gunicorn==20.1.0",
    "jinja2==3.1.3",
    "joblib==1.3.2",
    "jsonschema==3.2.0",
    "kubernetes>=34.1.0",
    "lightgbm>=4.6.0",
    "mlflow==2.11.2",
    "numpy==1.26.4",
    "pandas==2.2.1",
    "pyyaml==6.0.1",
    "querystring-parser==1.2.4",
    "requests==2.31.0",
    "scikit-learn==1.4.1.post1",
    "scipy==1.12.0",
    "seldon-core==1.18.2",
    "torch>=2.9.0",
]
